refactor: Simplify action builder code.
