--- src/components/Guilds/ActionsBuilder/Option/styles.tsx
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-import { Box } from 'components/Guilds/common/Layout';
-
-export const OptionWrapper = styled(Box)`
-  padding: 1rem;
-`;
-
-export const DetailWrapper = styled(Box)`
-  padding: 0.5rem 0;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-
-export const Detail = styled(Box)`
-  display: inline-flex;
-  margin-right: 0.75rem;
-`;
