+++ src/components/Guilds/ActionsBuilder/index.tsx
@@ -3,7 +3,7 @@ import { Header as CardHeader } from '../common/Card';
 import SidebarCard, {
   SidebarCardHeaderSpaced,
 } from 'components/Guilds/SidebarCard';
-import EditMode from './EditMode';
+import OptionsList from './OptionsList';
 import { Option } from './types';
 import { bulkEncodeCallsFromOptions } from 'hooks/Guilds/contracts/useEncodedCall';
 import EditButton from './common/EditButton';
@@ -34,16 +34,22 @@ export const ActionsBuilder: React.FC<ActionsBuilderProps> = ({
       header={
         <SidebarCardHeaderSpaced>
           <CardHeader>Actions</CardHeader>
-          <EditButton
-            variant="secondary"
-            onClick={() => (isEditable ? onSave() : onEdit())}
-          >
-            {isEditable ? 'Save' : 'Edit'}
-          </EditButton>
+          {editable && (
+            <EditButton
+              variant="secondary"
+              onClick={() => (isEditable ? onSave() : onEdit())}
+            >
+              {isEditable ? 'Save' : 'Edit'}
+            </EditButton>
+          )}
         </SidebarCardHeaderSpaced>
       }
     >
-      <EditMode isEditable={isEditable} options={options} onChange={onChange} />
+      <OptionsList
+        isEditable={isEditable}
+        options={options}
+        onChange={onChange}
+      />
     </SidebarCard>
   );
 };
